# Conflicts:
#	vue/src/components/BoardSidebar.vue

+++ sails/api/controllers/boards/delete.js
@@ -20,6 +20,10 @@ module.exports = {
       description: 'Missing parameters',
       statusCode: 400
     },
+    cannotDeleteDefault: {
+      description: 'Default board cannot be deleted',
+      statusCode: 409
+    },
     nonExistent: {
       description: 'Board does not exist in database',
       statusCode: 404
@@ -30,6 +34,11 @@ module.exports = {
     if(!inputs.boardId) {
       return exits.missingParams();
     }
+    var isDefaultBoard = await Board.findOne({id: inputs.boardId});
+    if(isDefaultBoard.boardType === 0) {
+      return exits.cannotDeleteDefault('Default Board may not be deleted');
+    }
+
     sails.log.info('BOARD_DELETE::: Trying to delete board ' + inputs.boardId);
     var deletedBoard = await Board.destroyOne({id: inputs.boardId});
     if(!deletedBoard) {
